added Volume<T>::operator() with long and size_t args, *not* using blitz accessors, because blitz only takes ints as index args and thus only limits to 32 bits access. Fixes: #5529
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/trunk@49925 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ aimsalgo/src/AimsZCat/main.cc
@@ -145,7 +145,7 @@ bool doitMesh( Process & p, const string &, Finder & )
 
 template<class T> bool ZCat::cat( AimsData<T> & out )
 {
-  int x, y, z, t, Z = 0;  
+  long x, y, z, t, Z = 0;
   list<string>::const_iterator	it;
   unsigned n = 0;
   bool copy_header = true;
